nfn data structure
